Fix collection popup window - PMT #104308
Firefox was throwing a "permission denied" error when modifying
the document of a window opened from the content script with
`window.open()`. It seems like Firefox is stricter than Chrome
or Safari about this -
https://developer.mozilla.org/en-US/docs/Web/Security/Same-origin_policy#Cross-origin_script_API_access

I've refactored the code to use a "Panel" from the firefox extension
API.
https://developer.mozilla.org/en-US/Add-ons/SDK/High-Level_APIs/panel

This commit contains some changes to the common code, which I'll have to
refactor out so the common code continues to work on all browsers.